+++ library/src/actions/rfcEmail/rfcEmail.test.ts
@@ -71,7 +71,7 @@ describe('rfcEmail', () => {
     });
 
     test('for simple email', () => {
-      expectNoActionIssue(action, ['email@example.com']);
+      expectNoActionIssue(action, ['email@example.com', 'Email@Example.COM']);
     });
 
     test('for very short email', () => {
@@ -312,7 +312,12 @@ describe('rfcEmail', () => {
         'あいうえお@example.com',
         'email@あいうえお.com',
         'email@example.あいう',
+        'K@example.com',
+        'ſ@example.com',
         'Kſ@example.com',
+        'email@K.com',
+        'email@ſ.com',
+        'email@example.Kſ',
       ]);
     });